+++ src/auth/passthrough.ts
@@ -13,8 +13,7 @@
 // limitations under the License.
 
 import {GaxiosOptions} from 'gaxios';
-import {AuthClient} from './authclient';
-import {GetAccessTokenResponse, Headers} from './oauth2client';
+import {AuthClient, GetAccessTokenResponse, Headers} from './authclient';
 
 /**
  * An AuthClient without any Authentication information. Useful for: